chore: refactor autosuggest layers
